requirements.txt
wandb
torch
pytorch-lightning
hydra-core
scikit-learn
einops
easydict
panda3d-simplepbr==0.10
h5py

## New dependencies
pandarallel
pandas
pydantic
rich
devtools
